fix: type errors
