+++ apps/frontend/src/components/date/datetime.tsx
@@ -30,11 +30,11 @@ function Component({
   children,
   ...props
 }: IProps) {
-  const clientInfo = useClient();
+  const client = useClient();
   const [isCopied, switchCopiedStatus] = useState(false);
-  const formattedDateTime = !clientInfo.isClient
+  const formattedDateTime = !client
     ? dateTime
-    : formatDateTime(clientInfo.locale, dateTime);
+    : formatDateTime(client.locale, dateTime);
 
   return (
     <div {...props}>
