+++ apps/frontend/src/entities/task/stats.tsx
@@ -10,7 +10,7 @@ import type { ILocalizableProps, ITranslatableProps } from "#components/types";
 import type { IPlace } from "#entities/place";
 
 interface IProps extends ILocalizableProps, ITranslatableProps {
-  translation: ILocalization["stats_tasks"];
+  translation: ILocalization["pages"]["stats_tasks"];
   placeID?: IPlace["id"];
 }
