+++ apps/frontend/src/pages/[lang]/account/index.module.scss
