+++ apps/frontend/src/pages/[lang]/stats/tasks/index.tsx
@@ -25,7 +25,7 @@ function TasksStatsPage({
             <TasksStats
               language={lang}
               commonTranslation={common}
-              translation={translation}
+              translation={t}
             />
           </OverviewHeader>
         )}
